hint on how to structure
